feat(readme.ts): add new helper function `readmeText` to generate readme text based on template file and plugin metadata
The `readmeText` function takes in several parameters: `templateFile`, `mainFile`, `packageFile`, `changelog`, and `changelogLength`. It imports the `fileContents` and `log` functions from the `utils` module, as well as the `parsePluginMetadata` and `formatChangelog` functions from the `meta` and `changelog` modules respectively.

Inside the function, it parses the plugin metadata using the `parsePluginMetadata` function and assigns the formatted changelog to the `PLUGIN_CHANGELOG` property of the metadata object.

If the `debug` environment variable is set, it logs the metadata object using the `log` function.

Next, it reads the contents of the template file using the `fileContents` function and assigns it to the `readme` variable. It then replaces variables in the template with their corresponding values from the metadata object using the `replace` method.

Finally, it returns the generated readme text.